Move custom handlers to separate files
