+++ index.js
@@ -1,1 +1,1 @@
-!function(e,t){"object"==typeof exports&&"object"==typeof module?module.exports=t():"function"==typeof define&&define.amd?define([],t):"object"==typeof exports?exports.SelectMultiple=t():e.SelectMultiple=t()}(self,(()=>(()=>{"use strict";var e={586:(e,t,n)=>{n.d(t,{Z:()=>c});var l=n(81),o=n.n(l),i=n(645),s=n.n(i)()(o());s.push([e.id,'/**\n * All the necessary styles for a select-multiple package\n */\n\nselect-multiple *, select-multiple *::before, select-multiple *::after {\n  box-sizing: border-box;\n}\n\nselect-multiple {\n  position: relative;\n  display: inline-block;\n\n  --select-multiple-font-color: #000000;\n  --select-multiple-font-invert-color: #000000;\n  --select-multiple-background-color: #ffffff;\n  --select-multiple-border-color: #767676;\n  --select-multiple-shadow-color: #00000040;\n\n  --select-multiple-font-size: 13.3333px;\n\n  --select-multiple-tag-height: 20px;\n  --select-multiple-tag-gap: 4px;\n\n  --select-multiple-list-height: 206px;\n  --select-multiple-item-height: 17px;\n}\n\n@media (prefers-color-scheme: dark) {\n  select-multiple {\n    --select-multiple-font-color: #ffffff;\n    --select-multiple-font-invert-color: #000000;\n    --select-multiple-background-color: #3b3b3b;\n    --select-multiple-border-color: #858585;\n  }\n}\n\nselect-multiple + select[multiple] {\n  display: none !important;\n}\n\nselect-multiple[data-loading="true"] *:not(select-input) {\n  visibility: hidden;\n  opacity: 0;\n}\n\nselect-multiple select-input {\n  display: flex;\n  flex-wrap: wrap;\n  gap: var(--select-multiple-tag-gap);\n  background: var(--select-multiple-background-color);\n  border: 1px solid var(--select-multiple-border-color);\n  border-radius: 2px;\n  min-height: 30px;\n  padding: 4px;\n}\n\nselect-multiple select-input::before {\n  content: attr(data-placeholder) "\\200B";\n  color: var(--select-multiple-font-color);\n  line-height: var(--select-multiple-tag-height);\n  font-size: var(--select-multiple-font-size);\n  font-family: Arial;\n  opacity: 0.5;\n}\n\nselect-multiple select-input:not(:empty)::before {\n  content: "\\200B"; /* Zero-width space */\n}\n\nselect-multiple select-input select-tag:first-child {\n  margin-left: calc(-1 * var(--select-multiple-tag-gap));\n}\n\nselect-multiple select-input select-tag {\n  display: flex;\n  align-items: center;\n  column-gap: 2px;\n  font-family: Arial;\n  font-size: var(--select-multiple-font-size);\n  color: var(--select-multiple-font-color);\n  height: var(--select-multiple-tag-height);\n  background: var(--select-multiple-background-color);\n  border: 1px solid var(--select-multiple-border-color);\n  border-radius: 2px;\n  padding: 0 3px;\n  cursor: default;\n}\n\nselect-multiple select-input select-tag select-remove {\n  font-size: var(--select-multiple-tag-height);\n  line-height: var(--select-multiple-tag-height);\n  user-select: none;\n  cursor: pointer;\n}\n\nselect-multiple select-list {\n  visibility: hidden;\n  position: absolute;\n  list-style: none;\n  user-select: none;\n  overflow-y: scroll;\n  background: var(--select-multiple-background-color);\n  box-shadow: 0 0 16px var(--select-multiple-shadow-color);\n  border: 1px solid var(--select-multiple-border-color);\n  max-height: var(--select-multiple-list-height);\n  min-width: min-content;\n  width: 100%;\n  padding: 0;\n  margin: 0;\n  z-index: 1;\n}\n\nselect-multiple select-list[data-visible="true"] {\n  visibility: visible;\n}\n\nselect-multiple select-list select-item {\n  display: list-item;\n  font-family: Arial;\n  font-size: var(--select-multiple-font-size);\n  color: var(--select-multiple-font-color);\n  min-height: var(--select-multiple-item-height);\n  padding: 1px 4px;\n}\n\nselect-multiple select-list select-item[data-visible="false"] {\n  display: none;\n}\n\nselect-multiple select-list select-item:hover {\n  color: var(--select-multiple-font-invert-color);\n  background: Highlight;\n  cursor: pointer;\n}\n',""]);const c=s.toString()},828:(e,t,n)=>{n.d(t,{Z:()=>c});var l=n(81),o=n.n(l),i=n(645),s=n.n(i)()(o());s.push([e.id,"select-multiple {\n  --select-multiple-custom-font-color: #ffffff;\n  --select-multiple-custom-background-color: #5897fb;\n}\n\nselect-input {\n  min-height: 38px;\n  padding: 6px;\n  gap: 6px;\n}\n\nselect-input::before {\n  line-height: 24px;\n}\n\nselect-tag {\n  color: var(--select-multiple-custom-font-color);\n  background: var(--select-multiple-custom-background-color);\n  padding: 0 6px;\n  height: 24px;\n  border: none;\n}\n\nselect-tag:first-child {\n  margin-left: -6px;\n}\n\nselect-remove {\n  font-size: 24px;\n  line-height: 24px;\n}\n\nselect-item:hover {\n  color: var(--select-multiple-custom-font-color);\n  background: var(--select-multiple-custom-background-color);\n}\n",""]);const c=s.toString()},645:e=>{e.exports=function(e){var t=[];return t.toString=function(){return this.map((function(t){var n="",l=void 0!==t[5];return t[4]&&(n+="@supports (".concat(t[4],") {")),t[2]&&(n+="@media ".concat(t[2]," {")),l&&(n+="@layer".concat(t[5].length>0?" ".concat(t[5]):""," {")),n+=e(t),l&&(n+="}"),t[2]&&(n+="}"),t[4]&&(n+="}"),n})).join("")},t.i=function(e,n,l,o,i){"string"==typeof e&&(e=[[null,e,void 0]]);var s={};if(l)for(var c=0;c<this.length;c++){var E=this[c][0];null!=E&&(s[E]=!0)}for(var r=0;r<e.length;r++){var T=[].concat(e[r]);l&&s[T[0]]||(void 0!==i&&(void 0===T[5]||(T[1]="@layer".concat(T[5].length>0?" ".concat(T[5]):""," {").concat(T[1],"}")),T[5]=i),n&&(T[2]?(T[1]="@media ".concat(T[2]," {").concat(T[1],"}"),T[2]=n):T[2]=n),o&&(T[4]?(T[1]="@supports (".concat(T[4],") {").concat(T[1],"}"),T[4]=o):T[4]="".concat(o)),t.push(T))}},t}},81:e=>{e.exports=function(e){return e[1]}}},t={};function n(l){var o=t[l];if(void 0!==o)return o.exports;var i=t[l]={id:l,exports:{}};return e[l](i,i.exports,n),i.exports}n.n=e=>{var t=e&&e.__esModule?()=>e.default:()=>e;return n.d(t,{a:t}),t},n.d=(e,t)=>{for(var l in t)n.o(t,l)&&!n.o(e,l)&&Object.defineProperty(e,l,{enumerable:!0,get:t[l]})},n.o=(e,t)=>Object.prototype.hasOwnProperty.call(e,t);var l={};return(()=>{n.d(l,{default:()=>a});const e={ELEMENT_SELECT_BASE:"select",ELEMENT_SELECT_MULTIPLE:"select-multiple",ELEMENT_SELECT_INPUT:"select-input",ELEMENT_SELECT_LIST:"select-list",ELEMENT_SELECT_ITEM:"select-item",ELEMENT_SELECT_TAG:"select-tag",ELEMENT_SELECT_REMOVE:"select-remove",ATTRIBUTES_COMMON:"common",ATTRIBUTE_ROLE:"role",ATTRIBUTE_VALUE:"data-value",ATTRIBUTE_VISIBILITY:"data-visible",ATTRIBUTE_LOADING:"data-loading",ATTRIBUTE_PLACEHOLDER:"data-placeholder",ATTRIBUTE_TABINDEX:"tabindex",ATTRIBUTE_TEXT:"text",ATTRIBUTE_HTML:"html",THEME_MODERN:"modern"},t={[e.ELEMENT_SELECT_MULTIPLE]:{[e.ELEMENT_SELECT_INPUT]:{},[e.ELEMENT_SELECT_LIST]:{[e.ELEMENT_SELECT_ITEM]:{}}}},o={[e.ELEMENT_SELECT_MULTIPLE]:{[e.ATTRIBUTES_COMMON]:[{name:e.ATTRIBUTE_LOADING,value:!0}]},[e.ELEMENT_SELECT_INPUT]:{[e.ATTRIBUTES_COMMON]:[{name:e.ATTRIBUTE_ROLE,value:"input"},{name:e.ATTRIBUTE_TABINDEX,value:0}]},[e.ELEMENT_SELECT_LIST]:{[e.ATTRIBUTES_COMMON]:[{name:e.ATTRIBUTE_ROLE,value:"listbox"},{name:e.ATTRIBUTE_VISIBILITY,value:!1}]},[e.ELEMENT_SELECT_ITEM]:{[e.ATTRIBUTES_COMMON]:[{name:e.ATTRIBUTE_ROLE,value:"option"},{name:e.ATTRIBUTE_VISIBILITY,value:!0}]},[e.ELEMENT_SELECT_TAG]:{},[e.ELEMENT_SELECT_REMOVE]:{}};class SelectElement{constructor(t,n){if(this.element=t,n[e.ATTRIBUTE_HTML]&&(t.innerHTML=n[e.ATTRIBUTE_HTML]),n[e.ATTRIBUTE_TEXT]&&(t.innerText=n[e.ATTRIBUTE_TEXT]),n[e.ATTRIBUTES_COMMON])for(const l of n[e.ATTRIBUTES_COMMON])t.setAttribute(l.name,l.value)}visibility(t){this.element.setAttribute(e.ATTRIBUTE_VISIBILITY,t)}loading(t){this.element.setAttribute(e.ATTRIBUTE_LOADING,t)}remove(){this.element.remove()}}class SelectItem extends SelectElement{constructor(t){super(document.createElement(e.ELEMENT_SELECT_ITEM),t)}}class SelectTag extends SelectElement{constructor(t){super(document.createElement(e.ELEMENT_SELECT_TAG),t)}}class SelectRemove extends SelectElement{constructor(t){super(document.createElement(e.ELEMENT_SELECT_REMOVE),t)}}const i={[e.ELEMENT_SELECT_MULTIPLE]:class SelectMultiple extends SelectElement{constructor(t){super(document.createElement(e.ELEMENT_SELECT_MULTIPLE),t)}},[e.ELEMENT_SELECT_INPUT]:class SelectInput extends SelectElement{constructor(t){super(document.createElement(e.ELEMENT_SELECT_INPUT),t)}},[e.ELEMENT_SELECT_LIST]:class SelectList extends SelectElement{constructor(t){super(document.createElement(e.ELEMENT_SELECT_LIST),t)}},[e.ELEMENT_SELECT_ITEM]:SelectItem,[e.ELEMENT_SELECT_TAG]:SelectTag,[e.ELEMENT_SELECT_REMOVE]:SelectRemove},s={[e.ELEMENT_SELECT_MULTIPLE]:(t,n)=>{const l=t[e.ELEMENT_SELECT_BASE],o=t[e.ELEMENT_SELECT_MULTIPLE];l.before(o.element)},[e.ELEMENT_SELECT_INPUT]:(t,n)=>{const l=t[e.ELEMENT_SELECT_BASE],o=t[e.ELEMENT_SELECT_MULTIPLE],i=t[e.ELEMENT_SELECT_INPUT];o.element.append(i.element);const s=l.getAttribute("placeholder");s&&i.element.setAttribute(e.ATTRIBUTE_PLACEHOLDER,s)},[e.ELEMENT_SELECT_LIST]:(t,n)=>{const l=t[e.ELEMENT_SELECT_MULTIPLE],o=t[e.ELEMENT_SELECT_LIST];l.element.append(o.element),l.element.addEventListener("click",(()=>{o.visibility(!0)})),document.addEventListener("click",(e=>{l.element.contains(e.target)||o.visibility(!1)}))},[e.ELEMENT_SELECT_ITEM]:(t,n)=>{const l=t[e.ELEMENT_SELECT_BASE],i=t[e.ELEMENT_SELECT_MULTIPLE],s=t[e.ELEMENT_SELECT_INPUT],c=t[e.ELEMENT_SELECT_LIST],E=l.querySelectorAll("option");let r="";for(const t of E){const n=o[e.ELEMENT_SELECT_ITEM];n[e.ATTRIBUTES_COMMON].push({name:e.ATTRIBUTE_VALUE,value:t.value}),n[e.ATTRIBUTE_TEXT]=t.innerText,t.innerText.length>r.length&&(r=t.innerText);const l=new SelectItem(n);c.element.append(l.element),l.element.addEventListener("click",(n=>{l.visibility(!1),t.setAttribute("selected","selected");const i=o[e.ELEMENT_SELECT_TAG];i[e.ATTRIBUTE_TEXT]=n.target.innerText;const c=new SelectTag(i);s.element.append(c.element);const E=o[e.ELEMENT_SELECT_REMOVE];E[e.ATTRIBUTE_HTML]="&times;";const r=new SelectRemove(E);c.element.append(r.element),r.element.addEventListener("click",(()=>{l.visibility(!0),t.removeAttribute("selected"),c.remove()}),{once:!0})}))}if(n.fullwidth&&(i.element.style.width="100%"),!n.fullwidth){const{width:t}=c.element.getBoundingClientRect(),n=o[e.ELEMENT_SELECT_TAG];n[e.ATTRIBUTE_TEXT]=r;const l=new SelectTag(n);s.element.append(l.element);const i=o[e.ELEMENT_SELECT_REMOVE];i[e.ATTRIBUTE_HTML]="&times;";const E=new SelectRemove(i);l.element.append(E.element);const{width:T}=c.element.getBoundingClientRect(),a=t>T?t:T;s.element.style.width=`${a}px`,l.remove()}i.loading(!1)}};class SelectBuilder{constructor(e,t){if(!(e instanceof HTMLSelectElement))throw new TypeError("Provided element is not an instance of HTMLSelectElement");this.config=t,this.select=e,this.wrapper=null}buildOnTopOfSelect(){const n={},l=[];n[e.ELEMENT_SELECT_BASE]=this.select,this.createChildElements(l,n,t);for(const e of l)this.customElementHandler(e,n)}removeBuiltSelect(){this.wrapper.remove()}createChildElements(e,t,n){for(const[l,c]of Object.entries(n)){const n=new i[l](o[l]);t[l]=n,null===this.wrapper&&(this.wrapper=n),s[l]instanceof Function&&e.push(s[l]),this.createChildElements(e,t,c)}}customElementHandler(e,t){e(t,this.config)}}var c=n(828);const E={[e.THEME_MODERN]:c.Z},r=e=>{const t=document.querySelector("head"),n=document.createElement("style");return n.innerHTML=e,t.append(n),n};class ConfigParser{constructor(e){this.config=e}parseConfigOptions(){return this.config&&this.config.theme&&this.config.theme.name&&!this.config.theme.styles&&E[this.config.theme.name]&&(this.config.theme.styles=E[this.config.theme.name]),this.config}}class StyleParser{constructor(e,t){this.core=e,this.theme=t,this.coreElement=null,this.themeElement=null}mountCoreStylesheet(){this.coreElement=r(this.core)}mountThemeStylesheet(){this.themeElement=r(this.theme)}removeOverlapingStyles(){for(const e of this.coreElement.sheet.cssRules)for(const t of this.themeElement.sheet.cssRules)if(e.selectorText&&e.selectorText.endsWith(t.selectorText))for(const n of t.style)e.style.removeProperty(n)}}var T=n(586);const a=class src_SelectMultiple{static THEME_MODERN=e.THEME_MODERN;constructor(e){const t=new ConfigParser(e);this.config=t.parseConfigOptions();const n=this.config?.theme?.styles,l=new StyleParser(T.Z,n);l.mountCoreStylesheet(),l.theme&&(l.mountThemeStylesheet(),l.removeOverlapingStyles()),this.__builders__=[]}transform(e){const t=new SelectBuilder(e,this.config);this.__builders__.push({element:e,builder:t}),t.buildOnTopOfSelect()}transformAllSelects(e){const t=e||"select[multiple]",n=document.querySelectorAll(t);for(const e of n){const t=new SelectBuilder(e,this.config);this.__builders__.push({element:e,builder:t}),t.buildOnTopOfSelect()}}destroy(e){const t=this.__builders__.find((t=>t.element===e));t.builder instanceof SelectBuilder&&t.builder.removeBuiltSelect()}destroyAllSelects(e){const t=e||"select[multiple]",n=document.querySelectorAll(t);for(const e of n){const t=this.__builders__.find((t=>t.element===e));t.builder instanceof SelectBuilder&&t.builder.removeBuiltSelect()}}}})(),l=l.default})()));
+!function(e,t){"object"==typeof exports&&"object"==typeof module?module.exports=t():"function"==typeof define&&define.amd?define([],t):"object"==typeof exports?exports.SelectMultiple=t():e.SelectMultiple=t()}(self,(()=>(()=>{"use strict";var e={586:(e,t,n)=>{n.d(t,{Z:()=>s});var l=n(81),o=n.n(l),E=n(645),i=n.n(E)()(o());i.push([e.id,'/**\n * All the necessary styles for a select-multiple package\n */\n\nselect-multiple *, select-multiple *::before, select-multiple *::after {\n  box-sizing: border-box;\n}\n\nselect-multiple {\n  position: relative;\n  display: inline-block;\n\n  --select-multiple-font-color: #000000;\n  --select-multiple-font-invert-color: #000000;\n  --select-multiple-background-color: #ffffff;\n  --select-multiple-border-color: #767676;\n  --select-multiple-shadow-color: #00000040;\n\n  --select-multiple-font-size: 13.3333px;\n\n  --select-multiple-tag-height: 20px;\n  --select-multiple-tag-gap: 4px;\n\n  --select-multiple-list-height: 206px;\n  --select-multiple-item-height: 17px;\n}\n\n@media (prefers-color-scheme: dark) {\n  select-multiple {\n    --select-multiple-font-color: #ffffff;\n    --select-multiple-font-invert-color: #000000;\n    --select-multiple-background-color: #3b3b3b;\n    --select-multiple-border-color: #858585;\n  }\n}\n\nselect-multiple + select[multiple] {\n  display: none !important;\n}\n\nselect-multiple[data-loading="true"] *:not(select-input) {\n  visibility: hidden;\n  opacity: 0;\n}\n\nselect-multiple select-input {\n  display: flex;\n  flex-wrap: wrap;\n  gap: var(--select-multiple-tag-gap);\n  background: var(--select-multiple-background-color);\n  border: 1px solid var(--select-multiple-border-color);\n  border-radius: 2px;\n  min-height: 30px;\n  padding: 4px;\n}\n\nselect-multiple select-input::before {\n  content: attr(data-placeholder) "\\200B";\n  color: var(--select-multiple-font-color);\n  line-height: var(--select-multiple-tag-height);\n  font-size: var(--select-multiple-font-size);\n  font-family: Arial;\n  opacity: 0.5;\n}\n\nselect-multiple select-input:not(:empty)::before {\n  content: "\\200B"; /* Zero-width space */\n}\n\nselect-multiple select-input select-tag:first-child {\n  margin-left: calc(-1 * var(--select-multiple-tag-gap));\n}\n\nselect-multiple select-input select-tag {\n  display: flex;\n  align-items: center;\n  column-gap: 2px;\n  font-family: Arial;\n  font-size: var(--select-multiple-font-size);\n  color: var(--select-multiple-font-color);\n  height: var(--select-multiple-tag-height);\n  background: var(--select-multiple-background-color);\n  border: 1px solid var(--select-multiple-border-color);\n  border-radius: 2px;\n  padding: 0 3px;\n  cursor: default;\n}\n\nselect-multiple select-input select-tag select-remove {\n  font-size: var(--select-multiple-tag-height);\n  line-height: var(--select-multiple-tag-height);\n  user-select: none;\n  cursor: pointer;\n}\n\nselect-multiple select-list {\n  visibility: hidden;\n  position: absolute;\n  list-style: none;\n  user-select: none;\n  overflow-y: scroll;\n  background: var(--select-multiple-background-color);\n  box-shadow: 0 0 16px var(--select-multiple-shadow-color);\n  border: 1px solid var(--select-multiple-border-color);\n  max-height: var(--select-multiple-list-height);\n  min-width: min-content;\n  width: 100%;\n  padding: 0;\n  margin: 0;\n  z-index: 1;\n}\n\nselect-multiple select-list[data-visible="true"] {\n  visibility: visible;\n}\n\nselect-multiple select-list select-item {\n  display: list-item;\n  font-family: Arial;\n  font-size: var(--select-multiple-font-size);\n  color: var(--select-multiple-font-color);\n  min-height: var(--select-multiple-item-height);\n  padding: 1px 4px;\n}\n\nselect-multiple select-list select-item[data-visible="false"] {\n  display: none;\n}\n\nselect-multiple select-list select-item:hover {\n  color: var(--select-multiple-font-invert-color);\n  background: Highlight;\n  cursor: pointer;\n}\n',""]);const s=i.toString()},828:(e,t,n)=>{n.d(t,{Z:()=>s});var l=n(81),o=n.n(l),E=n(645),i=n.n(E)()(o());i.push([e.id,"select-multiple {\n  --select-multiple-custom-font-color: #ffffff;\n  --select-multiple-custom-background-color: #5897fb;\n}\n\nselect-input {\n  min-height: 38px;\n  padding: 6px;\n  gap: 6px;\n}\n\nselect-input::before {\n  line-height: 24px;\n}\n\nselect-tag {\n  color: var(--select-multiple-custom-font-color);\n  background: var(--select-multiple-custom-background-color);\n  padding: 0 6px;\n  height: 24px;\n  border: none;\n}\n\nselect-tag:first-child {\n  margin-left: -6px;\n}\n\nselect-remove {\n  font-size: 24px;\n  line-height: 24px;\n}\n\nselect-item:hover {\n  color: var(--select-multiple-custom-font-color);\n  background: var(--select-multiple-custom-background-color);\n}\n",""]);const s=i.toString()},645:e=>{e.exports=function(e){var t=[];return t.toString=function(){return this.map((function(t){var n="",l=void 0!==t[5];return t[4]&&(n+="@supports (".concat(t[4],") {")),t[2]&&(n+="@media ".concat(t[2]," {")),l&&(n+="@layer".concat(t[5].length>0?" ".concat(t[5]):""," {")),n+=e(t),l&&(n+="}"),t[2]&&(n+="}"),t[4]&&(n+="}"),n})).join("")},t.i=function(e,n,l,o,E){"string"==typeof e&&(e=[[null,e,void 0]]);var i={};if(l)for(var s=0;s<this.length;s++){var c=this[s][0];null!=c&&(i[c]=!0)}for(var r=0;r<e.length;r++){var T=[].concat(e[r]);l&&i[T[0]]||(void 0!==E&&(void 0===T[5]||(T[1]="@layer".concat(T[5].length>0?" ".concat(T[5]):""," {").concat(T[1],"}")),T[5]=E),n&&(T[2]?(T[1]="@media ".concat(T[2]," {").concat(T[1],"}"),T[2]=n):T[2]=n),o&&(T[4]?(T[1]="@supports (".concat(T[4],") {").concat(T[1],"}"),T[4]=o):T[4]="".concat(o)),t.push(T))}},t}},81:e=>{e.exports=function(e){return e[1]}}},t={};function n(l){var o=t[l];if(void 0!==o)return o.exports;var E=t[l]={id:l,exports:{}};return e[l](E,E.exports,n),E.exports}n.n=e=>{var t=e&&e.__esModule?()=>e.default:()=>e;return n.d(t,{a:t}),t},n.d=(e,t)=>{for(var l in t)n.o(t,l)&&!n.o(e,l)&&Object.defineProperty(e,l,{enumerable:!0,get:t[l]})},n.o=(e,t)=>Object.prototype.hasOwnProperty.call(e,t);var l={};return(()=>{n.d(l,{default:()=>a});const e={ELEMENT_SELECT_BASE:"select",ELEMENT_SELECT_MULTIPLE:"select-multiple",ELEMENT_SELECT_INPUT:"select-input",ELEMENT_SELECT_LIST:"select-list",ELEMENT_SELECT_ITEM:"select-item",ELEMENT_SELECT_TAG:"select-tag",ELEMENT_SELECT_REMOVE:"select-remove",ATTRIBUTES_COMMON:"common",ATTRIBUTE_ROLE:"role",ATTRIBUTE_VALUE:"data-value",ATTRIBUTE_VISIBILITY:"data-visible",ATTRIBUTE_LOADING:"data-loading",ATTRIBUTE_PLACEHOLDER:"data-placeholder",ATTRIBUTE_TABINDEX:"tabindex",ATTRIBUTE_TEXT:"text",ATTRIBUTE_HTML:"html",THEME_MODERN:"modern"},t={[e.ELEMENT_SELECT_MULTIPLE]:{[e.ELEMENT_SELECT_INPUT]:{},[e.ELEMENT_SELECT_LIST]:{[e.ELEMENT_SELECT_ITEM]:{}}}},o={[e.ELEMENT_SELECT_MULTIPLE]:{[e.ATTRIBUTES_COMMON]:[{name:e.ATTRIBUTE_LOADING,value:!0}]},[e.ELEMENT_SELECT_INPUT]:{[e.ATTRIBUTES_COMMON]:[{name:e.ATTRIBUTE_ROLE,value:"input"},{name:e.ATTRIBUTE_TABINDEX,value:0}]},[e.ELEMENT_SELECT_LIST]:{[e.ATTRIBUTES_COMMON]:[{name:e.ATTRIBUTE_ROLE,value:"listbox"},{name:e.ATTRIBUTE_VISIBILITY,value:!1}]},[e.ELEMENT_SELECT_ITEM]:{[e.ATTRIBUTES_COMMON]:[{name:e.ATTRIBUTE_ROLE,value:"option"},{name:e.ATTRIBUTE_VISIBILITY,value:!0}]},[e.ELEMENT_SELECT_TAG]:{},[e.ELEMENT_SELECT_REMOVE]:{}};class SelectElement{constructor(t,n){if(this.element=t,n[e.ATTRIBUTE_HTML]&&(t.innerHTML=n[e.ATTRIBUTE_HTML]),n[e.ATTRIBUTE_TEXT]&&(t.innerText=n[e.ATTRIBUTE_TEXT]),n[e.ATTRIBUTES_COMMON])for(const l of n[e.ATTRIBUTES_COMMON])t.setAttribute(l.name,l.value)}visibility(t){this.element.setAttribute(e.ATTRIBUTE_VISIBILITY,t)}loading(t){this.element.setAttribute(e.ATTRIBUTE_LOADING,t)}remove(){this.element.remove()}}class SelectItem extends SelectElement{constructor(t){super(document.createElement(e.ELEMENT_SELECT_ITEM),t)}}class SelectTag extends SelectElement{constructor(t){super(document.createElement(e.ELEMENT_SELECT_TAG),t)}}class SelectRemove extends SelectElement{constructor(t){super(document.createElement(e.ELEMENT_SELECT_REMOVE),t)}}const E={[e.ELEMENT_SELECT_MULTIPLE]:class SelectMultiple extends SelectElement{constructor(t){super(document.createElement(e.ELEMENT_SELECT_MULTIPLE),t)}},[e.ELEMENT_SELECT_INPUT]:class SelectInput extends SelectElement{constructor(t){super(document.createElement(e.ELEMENT_SELECT_INPUT),t)}},[e.ELEMENT_SELECT_LIST]:class SelectList extends SelectElement{constructor(t){super(document.createElement(e.ELEMENT_SELECT_LIST),t)}},[e.ELEMENT_SELECT_ITEM]:SelectItem,[e.ELEMENT_SELECT_TAG]:SelectTag,[e.ELEMENT_SELECT_REMOVE]:SelectRemove},i={[e.ELEMENT_SELECT_MULTIPLE]:(t,n)=>{((t,n)=>{const l=t[e.ELEMENT_SELECT_BASE],o=t[e.ELEMENT_SELECT_MULTIPLE];l.before(o.element)})(t)},[e.ELEMENT_SELECT_INPUT]:(t,n)=>{((t,n)=>{const l=t[e.ELEMENT_SELECT_MULTIPLE],o=t[e.ELEMENT_SELECT_INPUT];l.element.append(o.element)})(t),((t,n)=>{const l=t[e.ELEMENT_SELECT_BASE],o=t[e.ELEMENT_SELECT_INPUT],E=l.getAttribute("placeholder");E&&o.element.setAttribute(e.ATTRIBUTE_PLACEHOLDER,E)})(t)},[e.ELEMENT_SELECT_LIST]:(t,n)=>{((t,n)=>{const l=t[e.ELEMENT_SELECT_MULTIPLE],o=t[e.ELEMENT_SELECT_LIST];l.element.append(o.element)})(t),((t,n)=>{const l=t[e.ELEMENT_SELECT_MULTIPLE],o=t[e.ELEMENT_SELECT_LIST];l.element.addEventListener("click",(()=>{o.visibility(!0)}))})(t),((t,n)=>{const l=t[e.ELEMENT_SELECT_MULTIPLE],o=t[e.ELEMENT_SELECT_LIST];document.addEventListener("click",(e=>{l.element.contains(e.target)||o.visibility(!1)}))})(t)},[e.ELEMENT_SELECT_ITEM]:(t,n)=>{((t,n)=>{const l=t[e.ELEMENT_SELECT_BASE],E=t[e.ELEMENT_SELECT_INPUT],i=t[e.ELEMENT_SELECT_LIST],s=l.querySelectorAll("option");for(const t of s){const n=o[e.ELEMENT_SELECT_ITEM];n[e.ATTRIBUTES_COMMON].push({name:e.ATTRIBUTE_VALUE,value:t.value}),n[e.ATTRIBUTE_TEXT]=t.innerText;const l=new SelectItem(n);i.element.append(l.element),l.element.addEventListener("click",(n=>{l.visibility(!1),t.setAttribute("selected","selected");const i=o[e.ELEMENT_SELECT_TAG];i[e.ATTRIBUTE_TEXT]=n.target.innerText;const s=new SelectTag(i);E.element.append(s.element);const c=o[e.ELEMENT_SELECT_REMOVE];c[e.ATTRIBUTE_HTML]="&times;";const r=new SelectRemove(c);s.element.append(r.element),r.element.addEventListener("click",(()=>{l.visibility(!0),t.removeAttribute("selected"),s.remove()}),{once:!0})}))}})(t),((t,n)=>{const l=t[e.ELEMENT_SELECT_BASE],E=t[e.ELEMENT_SELECT_MULTIPLE],i=t[e.ELEMENT_SELECT_INPUT],s=t[e.ELEMENT_SELECT_LIST];if(n.fullwidth)return void(E.element.style.width="100%");const c=l.querySelectorAll("option");let r="";for(const e of c)e.innerText.length>r.length&&(r=e.innerText);const{width:T}=s.element.getBoundingClientRect(),a=o[e.ELEMENT_SELECT_TAG];a[e.ATTRIBUTE_TEXT]=r;const u=new SelectTag(a);i.element.append(u.element);const m=o[e.ELEMENT_SELECT_REMOVE];m[e.ATTRIBUTE_HTML]="&times;";const p=new SelectRemove(m);u.element.append(p.element);const{width:L}=s.element.getBoundingClientRect(),d=T>L?T:L;i.element.style.width=`${d}px`,u.remove()})(t,n),((t,n)=>{t[e.ELEMENT_SELECT_MULTIPLE].loading(!1)})(t)}};class SelectBuilder{constructor(e,t){if(!(e instanceof HTMLSelectElement))throw new TypeError("Provided element is not an instance of HTMLSelectElement");this.config=t,this.select=e,this.wrapper=null}buildOnTopOfSelect(){const n={},l=[];n[e.ELEMENT_SELECT_BASE]=this.select,this.createChildElements(l,n,t);for(const e of l)this.customElementHandler(e,n)}removeBuiltSelect(){this.wrapper.remove()}createChildElements(e,t,n){for(const[l,s]of Object.entries(n)){const n=new E[l](o[l]);t[l]=n,null===this.wrapper&&(this.wrapper=n),i[l]instanceof Function&&e.push(i[l]),this.createChildElements(e,t,s)}}customElementHandler(e,t){e(t,this.config)}}var s=n(828);const c={[e.THEME_MODERN]:s.Z},r=e=>{const t=document.querySelector("head"),n=document.createElement("style");return n.innerHTML=e,t.append(n),n};class ConfigParser{constructor(e){this.config=e}parseConfigOptions(){return this.config&&this.config.theme&&this.config.theme.name&&!this.config.theme.styles&&c[this.config.theme.name]&&(this.config.theme.styles=c[this.config.theme.name]),this.config}}class StyleParser{constructor(e,t){this.core=e,this.theme=t,this.coreElement=null,this.themeElement=null}mountCoreStylesheet(){this.coreElement=r(this.core)}mountThemeStylesheet(){this.themeElement=r(this.theme)}removeOverlapingStyles(){for(const e of this.coreElement.sheet.cssRules)for(const t of this.themeElement.sheet.cssRules)if(e.selectorText&&e.selectorText.endsWith(t.selectorText))for(const n of t.style)e.style.removeProperty(n)}}var T=n(586);const a=class src_SelectMultiple{static THEME_MODERN=e.THEME_MODERN;constructor(e){const t=new ConfigParser(e);this.config=t.parseConfigOptions();const n=this.config?.theme?.styles,l=new StyleParser(T.Z,n);l.mountCoreStylesheet(),l.theme&&(l.mountThemeStylesheet(),l.removeOverlapingStyles()),this.__builders__=[]}transform(e){const t=new SelectBuilder(e,this.config);this.__builders__.push({element:e,builder:t}),t.buildOnTopOfSelect()}transformAllSelects(e){const t=e||"select[multiple]",n=document.querySelectorAll(t);for(const e of n){const t=new SelectBuilder(e,this.config);this.__builders__.push({element:e,builder:t}),t.buildOnTopOfSelect()}}destroy(e){const t=this.__builders__.find((t=>t.element===e));t.builder instanceof SelectBuilder&&t.builder.removeBuiltSelect()}destroyAllSelects(e){const t=e||"select[multiple]",n=document.querySelectorAll(t);for(const e of n){const t=this.__builders__.find((t=>t.element===e));t.builder instanceof SelectBuilder&&t.builder.removeBuiltSelect()}}}})(),l=l.default})()));
