+++ package.json
@@ -13,7 +13,6 @@
   },
   "devDependencies": {
     "css-loader": "^6.7.1",
-    "style-loader": "^3.3.1",
     "terser-webpack-plugin": "^5.3.1",
     "webpack": "^5.72.0",
     "webpack-cli": "^4.9.2"
